Implemented basic noise detection and monster investigation
The monster can detect noise, and investigate it. This is not based on audio, but rather actions the player takes and how far they are from the monster. Thus actions that make more noise will attract the monster from further away. The monster will go to the spot where it heard the noise coming from to investigate. But right now, it'll stay stuck there unless something else happens. So we need to get it to return to its routine if nothing happens.