HP, stamina, trap, and cooldown implemented
This has a stamina bar implemented in the UI that drains. It also has the three-hit system fully working, where the monster can damage them, but there's a 3 second window before the monster can hurt them again. And the player can take damage from traps too. Doesn't reset the scene if they die to the trap, but it does if they die to the monster.